backendskjema/requirements.txt
boto3==1.9.220
Django==2.2.2
django-admin-reset==0.3.1
django-cors-headers==3.0.2
django-crispy-forms==1.7.2
django-environ==0.4.5
django-filter==2.0.0
django-graphql-jwt==0.1.5
django-hardcopy==0.1.4
django-heroku==0.3.1
django-rest-knox==4.1.0
django-storages==1.7.1
django-tempus-dominus==5.1.2.5
django-webcam==0.1.4b2
djangorestframework==3.10.3
graphene==2.1.8
graphene-django==2.2.0
graphene-file-upload==1.2.2
graphql-core==2.2.1
graphql-relay==2.0.0
django-extensions==1.5.9
psycopg2==2.8.3
gunicorn==19.9.0
dj-database-url==0.5.0
whitenoise==4.1.4
dj-static==0.0.6
python-decouple==3.1
Pillow==6.1.0
drf-writable-nested==0.5.1
PyJWT==1.7.1
